requirements.txt
asgiref==3.8.1
Django==5.2.1
djangorestframework==3.16.0
pyspark==3.5.5
sqlparse==0.5.3
